svg/path/interfaces/SVGAnimatedPathData-removed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><metadata><h:link href="https://svgwg.org/svg2-draft/single-page.html#changes-paths" rel="help"/><h:link href="https://www.w3.org/TR/SVG11/paths.html#InterfaceSVGAnimatedPathData" rel="help"/><h:meta content="SVGAnimatedPathData interface is not supported." name="assert"/></metadata><path id="track" d="m 10 20 h 30"/><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script><![CDATA[
  test(function() {
    assert_true(window.SVGAnimatedPathData === undefined);

    var track = document.getElementById('track');
    assert_equals(track.__proto__, SVGPathElement.prototype);
    assert_true(track.pathSegList === undefined);
    assert_true(track.normalizedPathSegList === undefined);
    assert_true(track.animatedPathSegList === undefined);
    assert_true(track.animatedNormalizedPathSegList === undefined);
  });
  ]]></script></svg>
