LayoutTests/imported/w3c/web-platform-tests/svg/path/interfaces/SVGAnimatedPathData-removed-expected.txt


FAIL SVGAnimatedPathData-removed assert_equals: expected (undefined) undefined but got (object) object "[object SVGPathSegList]"
